feat: web 区域录屏
