+++ packages/recorder/src/main.ts
@@ -1,3 +1,2 @@
 import { audio, video, screen } from "../lib";
-
 console.log({ audio, video, screen });
