+++ packages/web/package.json
@@ -23,6 +23,7 @@
     "js-cookie": "^3.0.5",
     "plyr": "^3.7.8",
     "qr-scanner": "^1.4.2",
+    "re-resizable": "^6.9.11",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^13.2.1",
