+++ packages/web/src/pages/recorderAudio/index.module.scss
@@ -11,6 +11,9 @@
 		#recordings {
 			margin: 1rem 0;
 		}
+		.tools {
+			line-height: 34px;
+		}
 	}
 }
 .web {
