+++ packages/web/src/pages/recorderScreen/index.module copy.scss
@@ -1,0 +1,117 @@
+.recorderScreen {
+	width: 100vw;
+	height: 100vh;
+	overflow: hidden;
+	:global {
+		.content {
+			width: 100%;
+			height: calc(100% - 40px);
+			overflow: hidden;
+			border-bottom: 1px solid #ccc;
+			.tip {
+				position: absolute;
+				font-size: 34px;
+				z-index: 100;
+				background: #fff;
+				top: 0;
+				bottom: 40px;
+				left: 0;
+				right: 0;
+				text-align: center;
+				line-height: 300px;
+			}
+		}
+		.footer {
+			display: flex;
+			width: 100vw;
+			height: 40px;
+			align-items: center;
+			flex-direction: row-reverse;
+			overflow: hidden;
+			-webkit-app-region: drag;
+			background-color: rgba(255, 250, 250, 0.3);
+			.singleDigit {
+				font-size: 20px;
+				background-color: transparent;
+			}
+			.heightInput {
+				-webkit-app-region: no-drag;
+			}
+			.widthInput {
+				-webkit-app-region: no-drag;
+			}
+			.separtorContainer {
+				margin: 0 0 8px;
+				.separtor {
+					width: 4px;
+					height: 4px;
+				}
+			}
+			.recorderTools {
+				display: flex;
+				.playBtn {
+					color: #1677ff;
+					font-size: 30px;
+				}
+				.stopBtn {
+					color: red;
+					font-size: 30px;
+				}
+				.pauseBtn {
+					font-size: 30px;
+				}
+				.toggleMuteBtn {
+					font-size: 18px;
+				}
+				.muted {
+					color: rgba(255, 255, 255, 0.88);
+				}
+			}
+			.sizeIcon {
+				margin: 0 5px;
+			}
+			.drgan {
+				flex: auto;
+				height: 34px;
+			}
+			.toolbarIcon {
+				line-height: 30px;
+				width: 30px;
+				height: 30px;
+				margin-right: 10px;
+				padding: 0;
+				-webkit-app-region: no-drag;
+			}
+			.shotScreenBtn {
+				color: red;
+				cursor: pointer;
+			}
+			.recordIcon {
+				float: left;
+				margin: 0 10px 0 10px;
+				-webkit-app-region: no-drag;
+			}
+			.blink {
+				color: red;
+			}
+			.toolbarTitle {
+				line-height: 32px;
+				white-space: nowrap;
+			}
+			.formatSelect {
+				margin-right: 5px;
+			}
+			.widthInput {
+				margin-right: 5px;
+			}
+		}
+	}
+}
+
+.recorderScreen:hover {
+	:global {
+		.footer {
+			background-color: rgba(255, 255, 255, 1);
+		}
+	}
+}
