+++ packages/web/src/pages/recorderScreen/index.module.scss
@@ -3,107 +3,88 @@
 	height: 100vh;
 	overflow: hidden;
 	:global {
-		.content {
-			width: 100%;
-			height: calc(100% - 40px);
-			overflow: hidden;
-			border-bottom: 1px solid #ccc;
-			.tip {
-				position: absolute;
-				font-size: 34px;
-				z-index: 100;
-				background: #fff;
-				top: 0;
-				bottom: 40px;
-				left: 0;
-				right: 0;
-				text-align: center;
-				line-height: 300px;
+		.selectMedia {
+			position: absolute;
+			left: 50%;
+			transform: translateX(-50%);
+			margin-top: 50px;
+			z-index: 100;
+		}
+		.cropArea {
+			position: absolute;
+			z-index: 10;
+			outline: 2px solid #f0f0f0;
+			user-select: none;
+			.contentCropArea {
+				width: 100%;
+				height: 100%;
+			}
+			.innerCropArea {
+				height: 100%;
+				width: 100%;
+				cursor: move;
 			}
 		}
-		.footer {
+		.screenRecorder {
+			position: absolute;
+			z-index: 100;
+			outline: 2px solid #f0f0f0;
 			display: flex;
-			width: 100vw;
-			height: 40px;
-			align-items: center;
-			flex-direction: row-reverse;
-			overflow: hidden;
-			-webkit-app-region: drag;
-			background-color: rgba(255, 250, 250, 0.3);
-			.singleDigit {
-				font-size: 20px;
-				background-color: transparent;
-			}
-			.heightInput {
-				-webkit-app-region: no-drag;
-			}
-			.widthInput {
-				-webkit-app-region: no-drag;
-			}
-			.separtorContainer {
-				margin: 0 0 8px;
-				.separtor {
-					width: 4px;
-					height: 4px;
-				}
-			}
-			.recorderTools {
-				display: flex;
-				.playBtn {
-					color: #1677ff;
-					font-size: 30px;
-				}
-				.stopBtn {
-					color: red;
-					font-size: 30px;
-				}
-				.pauseBtn {
-					font-size: 30px;
-				}
-				.toggleMuteBtn {
-					font-size: 18px;
-				}
-				.muted {
-					color: rgba(255, 255, 255, 0.88);
-				}
-			}
-			.sizeIcon {
-				margin: 0 5px;
-			}
-			.drgan {
-				flex: auto;
-				height: 34px;
-			}
+			cursor: pointer;
+			background-color: #ccc;
 			.toolbarIcon {
+				font-size: 30px;
 				line-height: 30px;
-				width: 30px;
-				height: 30px;
-				margin-right: 10px;
-				padding: 0;
-				-webkit-app-region: no-drag;
-			}
-			.shotScreenBtn {
-				color: red;
-				cursor: pointer;
-			}
-			.recordIcon {
-				float: left;
-				margin: 0 10px 0 10px;
-				-webkit-app-region: no-drag;
 			}
 			.blink {
 				color: red;
 			}
-			.toolbarTitle {
-				line-height: 32px;
-				white-space: nowrap;
+			.recordBtn {
+				font-size: 18px;
 			}
-			.formatSelect {
-				margin-right: 5px;
+			.settingBtn {
+				line-height: 15px;
 			}
-			.widthInput {
-				margin-right: 5px;
+			.drgan {
+				flex: auto;
 			}
+			.playBtn {
+				color: #1677ff;
+			}
+			.stopBtn {
+				color: red;
+			}
+			.pauseBtn {
+				color: #000;
+			}
+			.toggleMuteBtn {
+				font-size: 18px;
+				color: #000;
+			}
+			.playRecorder {
+				line-height: 40px;
+				display: flex;
+				.toolbarTitle {
+					height: 30px;
+				}
+			}
+		}
+		.iframeRef {
+			position: absolute;
+			left: 0;
+			top: 0;
+		}
+		.videoRef {
+			width: 100%;
+			height: 100%;
+			position: absolute;
+			left: 0;
+			top: 0;
+		}
+		.control {
+			position: absolute;
+			right: 0;
+			bottom: 0;
 		}
 	}
 }
